moved comment_json to a lib folder
comment_json is a library to read json files that have comments in them. There is an existing comment json library but it uses a simple regex that fails on some of the minecraft json files.
This library first tries using the standard json library and then falls back to itself if that fails.