Separated model reading into two functions and caching raw models
Before if a model was used by multiple blockstates it was created from the JSON files each time. Now the base model is evaluated to a MinecraftMesh once and that value is returned each time.
Rotation as defined in the blockstate file is applied in the first function after loading the model either from JSON or using the cached version