Made blocks with textures but no elements show the missing cube
Some blocks that are hard coded have a texture defined but the model is hard coded. The previous code was displaying them like they would air but it was ambiguous that there was actually a block there. This will allow us to see where those blocks actually are and add models for them