Switched to using image paths rather than namespaced tuples
The old way of storing the namespace and relative path is very java centered and will cause issues implementing the Bedrock reader. They are now referred to by their file paths which is consistent for both.
The textures should be converted to the paths before creating the BlockMesh so that the data in the BlockMesh class does not rely on the resource pack.