Fixed writable directories
Previously the vanilla resource packs were downloaded to a location within the package and cache data was also saved within the package.
In some cases the install directory is not writable which caused the application to crash.
This adds a CACHE_DIR environment variable which can be set by the user or the application.
A default value is set if one is not set first. This uses platformdirs.
Moved the resource pack fixes up a directory.